pyproject.toml
[tool.poetry]
name = "xtl-read-assistant"
version = "0.0.2"
description = "X as a third language reading assistant"
authors = ["ffreemt"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/ffreemt/xtl-read-assistant"
repository = "https://github.com/ffreemt/xtl-read-assistant"
keywords = ["mt", "language", "learning"]
classifiers = [
    "Environment :: Console",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
include = [
    "LICENSE",
]

[tool.poetry.dependencies]
python = "^3.6"
pyperclip = "^1.7.0"
python-dotenv = "^0.12.0"
logzero = "^1.5.0"
environs = "^7.3.1"
pynput = "^1.6.8"
absl-py = "^0.9.0"
polyglot = "^16.7.4"
janus = "^0.4.0"
deepl-tr-async = "^0.0.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

# New scripts
[tool.poetry.scripts]
read-assist = "xtl_read_assistant.read_assist:main"
# read-assistant = "xtl_read_assistant.__main__:main"
# hotkey = "xtl_read_assistant.hotkey:main

# my-scri= "xtl_read_assistant.log_revision:start"
# hello = "xtl_read_assistant.read_assistant:hello"


[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
